Fix ZstdEncoder not producing data when source is smaller than block … (#15894)
…size (#15859)

Motivation:
If the size of the input buffer is smaller than the configured block
size (64 KB by default), then ZstdEncoder.write only produces an empty
buffer. With the default sizes, this makes the encoder unusable for
variable size content. This is for fixing the
issue(https://github.com/netty/netty/issues/15340)

Modification:

Return the uncompressed data if for small size data

Result:

Fixes #15340

Signed-off-by: xingrufei <qhdxssm@qq.com>
Co-authored-by: skyguard1 <qhdxssm@qq.com>