Fix MpscIntQueue bug (#16023) (#16027)
Motivation:
The int queue is always sized as a power of two, but the empty value was
only set for the specified capacity number of entries.

Modification:
Fill the full length of the underlying array with the empty entry value.

Result:
No surprise zero values when the empty value is different from zero.

Co-authored-by: Chris Vest <christianvest_hansen@apple.com>